#1426 Additional refactoring to improve the code structure.
