+++ platform-dao/src/main/java/ua/com/fielden/platform/entity/ReferenceHierarchyDao.java
@@ -3,17 +3,17 @@ package ua.com.fielden.platform.entity;
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.toList;
 import static ua.com.fielden.platform.entity.AbstractUnionEntity.unionProperties;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyActions.EDIT;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyActions.REFERENCE_HIERARCHY;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyLevel.REFERENCES;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyLevel.REFERENCE_BY_INSTANCE;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyLevel.REFERENCE_GROUP;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyLevel.TYPE;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.fetchKeyAndDescOnly;
 import static ua.com.fielden.platform.entity.query.metadata.DataDependencyQueriesGenerator.queryForDependentTypeDetails;
 import static ua.com.fielden.platform.entity.query.metadata.DataDependencyQueriesGenerator.queryForDependentTypesSummary;
 import static ua.com.fielden.platform.error.Result.failure;
 import static ua.com.fielden.platform.error.Result.failuref;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyActions.EDIT;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyActions.REFERENCE_HIERARCHY;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel.REFERENCES;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel.REFERENCE_BY_INSTANCE;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel.REFERENCE_GROUP;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel.TYPE;
 import static ua.com.fielden.platform.reflection.TitlesDescsGetter.getEntityTitleAndDesc;
 import static ua.com.fielden.platform.reflection.TitlesDescsGetter.getTitleAndDesc;
 import static ua.com.fielden.platform.utils.EntityUtils.hasDescProperty;
@@ -45,8 +45,8 @@ import ua.com.fielden.platform.entity.query.metadata.DataDependencyQueriesGenera
 import ua.com.fielden.platform.entity.query.model.AggregatedResultQueryModel;
 import ua.com.fielden.platform.error.Result;
 import ua.com.fielden.platform.pagination.IPage;
+import ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel;
 import ua.com.fielden.platform.reflection.Finder;
-import ua.com.fielden.platform.reflection.TitlesDescsGetter;
 import ua.com.fielden.platform.security.user.User;
 import ua.com.fielden.platform.utils.EntityUtils;
 import ua.com.fielden.platform.utils.Pair;
