+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/ref_hierarchy/ReferenceHierarchyProducer.java
@@ -1,7 +1,7 @@
-package ua.com.fielden.platform.entity;
+package ua.com.fielden.platform.web.ref_hierarchy;
 
 import static java.lang.String.format;
-import static ua.com.fielden.platform.entity.ReferenceHierarchyLevel.REFERENCE_INSTANCE;
+import static ua.com.fielden.platform.ref_hierarchy.ReferenceHierarchyLevel.REFERENCE_INSTANCE;
 import static ua.com.fielden.platform.entity.query.fluent.EntityQueryUtils.fetchKeyAndDescOnly;
 import static ua.com.fielden.platform.utils.EntityUtils.hasDescProperty;
 import static ua.com.fielden.platform.utils.EntityUtils.isPersistedEntityType;
@@ -11,6 +11,9 @@ import org.apache.commons.lang.StringUtils;
 
 import com.google.inject.Inject;
 
+import ua.com.fielden.platform.entity.AbstractEntity;
+import ua.com.fielden.platform.entity.DefaultEntityProducerWithContext;
+import ua.com.fielden.platform.entity.ReferenceHierarchy;
 import ua.com.fielden.platform.entity.factory.EntityFactory;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
 import ua.com.fielden.platform.entity.query.fluent.fetch;
