+++ platform-web-ui/src/main/java/ua/com/fielden/platform/web/ref_hierarchy/ReferenceHierarchyWebUiConfig.java
@@ -1,4 +1,4 @@
-package ua.com.fielden.platform.web.action;
+package ua.com.fielden.platform.web.ref_hierarchy;
 
 import static ua.com.fielden.platform.error.Result.failuref;
 import static ua.com.fielden.platform.web.centre.api.actions.impl.EntityActionBuilder.action;
@@ -7,7 +7,6 @@ import static ua.com.fielden.platform.web.centre.api.context.impl.EntityCentreCo
 import com.google.inject.Injector;
 
 import ua.com.fielden.platform.entity.ReferenceHierarchy;
-import ua.com.fielden.platform.entity.ReferenceHierarchyProducer;
 import ua.com.fielden.platform.web.centre.api.actions.EntityActionConfig;
 import ua.com.fielden.platform.web.centre.api.context.CentreContextConfig;
 import ua.com.fielden.platform.web.minijs.JsCode;
